Ball position and better collision
I cleaned up the code a bit, making it easier to read. I added all the balls for a standard 8 ball and spawned them into their correct position. I also moved the collision function into the ball class, cleaning up the code. The collisions worked most of the time, but there were some cases where the ball may phase through the other. This was due to two errors. One is that I was excluding stationary balls to increase performance, which is a mistake. I removed that line and no performance hit was noticed. The second error, was the check only happened once per code run (roughly 60 times per second) meaning that a ball moving faster than it's radius per tick could phase through. To fix this, I just increased the number of steps per code run, making the check more frequent and therefore accurate. I cleaned up the update function while I was at it, making it shorter and efficient now that collision was a method of any ball object.